Heroku: deploy with Maven plugin
Configure the POM for deploying to Heroku with the official maven
plugin. Authenticate with "heroku login" or set the HEROKU_API_KEY
environment variable, then run

    mvn heroku:deploy

Heroku must first be set to run a dedicated "stack", which amounts to a
pre-baked container image. This is most easily done with

    heroku stack:set heroku-18 --app crontention

but it is possible to use dedicated, more lightweight images.