Heroku: deploy with Docker JVM
Adjust the sample Quarkus JVM Dockerfile to use Java 11 and support
deployment to Heroku. Authenticate with "heroku login", then run

    ./mvnw package
    docker build \
        -f src/main/docker/Dockerfile.jvm \
        -t registry.heroku.com/crontention/web .
    heroku container:login
    docker push registry.heroku.com/crontention/web
    heroku container:release --app crontention web

In contrast to Maven deployment, the application does not roll out
without the explicit "container:release" step, but that step
automatically switches the Heroku "stack" as necessary (to "container").
To deploy with Maven later on, the stack must first be changed manually
with "stack:set".