Heroku: deploy with Docker native
Adjust the sample Quarkus native Dockerfile to support deployment to
Heroku, and the POM file to build in a container with a Java 11 enabled
GraalVM 20 instead of the default 19.3.1-java8. Authenticate with
"heroku login", then run

    ./mvnw package -Pnative
    docker build \
        -f src/main/docker/Dockerfile.native \
        -t registry.heroku.com/crontention/web .
    heroku container:login
    docker push registry.heroku.com/crontention/web
    heroku container:release --app crontention web

In contrast to Maven deployment, the application does not roll out
without the explicit "container:release" step, but that step
automatically switches the Heroku "stack" as necessary (to "container").
To deploy with Maven later on, the stack must first be changed manually
with "stack:set".