Indicate waiting for back-end in UI
Some pathological expressions can take a while to evaluate, during which
the UI appears unresponsive. Worse, the Heroku deployment will typically
be idling and need several seconds to awaken before any evaluation can
begin. When evaluating, disable interactivity and display a status, then
await a signal to restore interactivity and hide the status. Route all
evaluation through form submission to make this flow reliable.

When there is no input we don't need to await awakening before doing
nothing so we skip the fetch entirely. It's still helpful to communicate
that we reacted to the input and pretended to do work so disable
interactivity for just a little bit.