Support virtual command field
Quartz cron expressions are pure schedules, with no other identifying
information. This makes it difficult to reason about more than a few at
a time. Pretend they have an optional command field to the right of the
optional year field that users can disambiguate expressions with. This
field is more like a generic "name" or "meta" field but "command" is
analogous to the crontab syntax; and besides, a Quartz cron expression
doesn't exist without something to do, despite that not being embedded
in the expression syntax.