Clear heat map on invalid input
Loading a URL representing an invalid input leaves the heat map empty.
Submitting an invalid input after a valid input leaves the heat map as
it was after the valid input. This behaviour is inconsistent. The only
way to make it consistent is to clear the heat map on ajax evaluation
with invalid input.