Click rendered expression to highlight its cells
After evaluating expressions, replace the form with a rendered view of
the expressions. Each of these expression entries can be clicked to
highlight the heat map cells that expression activates in. This makes it
much easier to answer the question "when does this expression activate?"
and does not compromise the questions "how much activates when?" and
"what activates when?"

Clicking an expression will deactivate the highlighted expression and
highlight the newly clicked one. Clicking an already highlighted
expression will deactivate it. The contextually available Edit button
replaces the rendered expressions with the populated form.

This implementation generates an ID for each provided expression and
uses that as a CSS class later queried with querySelectorAll(). This
seems a bit clumsy, and the ID generation is finicky, but it works.