Replace HTML hidden attr with CSS class
The HTML "hidden" global attribute loses its effect in the presence of
any non-default CSS "display" property value [1][2]. This means we can't
manage visibility consistently if we need to hide any element with a
non-default "display" value. To retain consistent behaviour, replace the
attribute with a dedicated CSS class and introduce hide and show
functions that set and remove said class. This also eliminates the
awkward negation involved in using the "hidden" attribute.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden
[2] https://css-tricks.com/the-hidden-attribute-is-visibly-weak/