Explicitly track cells by expression
When redrawing cells, also store pointers to them in a global map
indexed by expressions. Reference that map instead of the
IDs-as-CSS-classes and remove the special CSS classes. I have no idea
how this affects performance relative to the CSS class approach but
there are no perceptible changes and the complex ID management is
replaced by comparatively simpler but equivalent state management.